PPs-62: Refactor Button atom
Problem:
* Button needs to be translated to atomic design

Solution:
* Refactor Button subcomponent as atom

Note:
* This introduced theme file
