+++ scripts/file_templates/componentStory.txt
@@ -1,7 +1,7 @@
 import React from 'react';
 import { withKnobs } from '@storybook/addon-knobs';
 
-import $component_name from '../index';
+import $component_name from '.';
 
 export default {
     title: '$component_types/$component_name',
