+++ scripts/file_templates/componentTest.txt
@@ -2,7 +2,7 @@ import React from 'react';
 import renderer from 'react-test-renderer';
 import { mount, shallow } from 'enzyme';
 
-import $component_name from '..';
+import $component_name from '.';
 
 describe('$component_name', () => {
     it('should shallowly render without crashing', () => {
