+++ src/components/atoms/Button/Button.stories.js
@@ -1,10 +1,12 @@
 import React from 'react';
+import { withKnobs } from '@storybook/addon-knobs';
 
-import Button from '../index';
+import Button from '.';
 
 export default {
-    title: 'Subcomponents/Button',
+    title: 'Atoms/Button',
     component: Button,
+    decorators: [withKnobs],
     argTypes: {
         
     },
@@ -14,5 +16,5 @@ const Template = (args) => <Button {...args} />;
 
 export const Primary = Template.bind({});
 Primary.args = {
-    text: 'Test Button',
+    text: 'Test Button'
 };
