+++ src/components/atoms/Button/Button.test.js
@@ -1,12 +1,15 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { shallow } from 'enzyme';
+import { mount, shallow } from 'enzyme';
 
-import Button from '..';
+import Button from '.';
 
 describe('Button', () => {
-    it('should render without crashing', () => {
+    it('should shallowly render without crashing', () => {
         shallow(<Button />);
+    });
+    it('should render without crashing', () => {
+        mount(<Button />);
     });
     it('should match snapshot', () => {
         const tree = renderer
