+++ src/components/styled-elements/index.js
@@ -1,0 +1,5 @@
+import StyledButton from './Button';
+
+export {
+    StyledButton,
+};
